Another pass on 'claiming' orders
