+++ webapp/src/components/UnclaimedOrdersWidget.js
@@ -3,6 +3,7 @@ import { useUser } from "../state/useUser";
 import { LayoutContainer } from "../state/withLayout";
 import React from "react";
 import Button from "react-bootstrap/Button";
+import {Link} from "react-router-dom";
 
 export default function UnclaimedOrdersWidget() {
   const { user } = useUser();
@@ -15,7 +16,8 @@ export default function UnclaimedOrdersWidget() {
         variant="outline-dark"
         size="sm"
         className="w-100 d-flex justify-content-between"
-        href="/order-history?unclaimed=true"
+        as={Link}
+        to="/unclaimed-orders"
       >
         {t("food:unclaimed_orders", { unclaimedOrdersCount: user.unclaimedOrdersCount })}
         <span className="ms-auto">
